[Refactor] 불필요한 쿼리 방지를 위해 ScheduledNotificationRepository.deleteAllByNotificationTypeInAndTargetUserIdIn()을 jpql로 작성

+++ caramel-domain/src/main/kotlin/com/whatever/caramel/domain/notification/service/ScheduledNotificationService.kt
@@ -33,7 +33,7 @@ class ScheduledNotificationService(
     fun deleteScheduledNotifications(
         targetUserIds: Set<Long>,
         notificationTypes: Set<NotificationType>,
-    ): Long {
+    ): Int {
         return scheduledNotificationRepository.deleteAllByNotificationTypeInAndTargetUserIdIn(
             notificationTypes = notificationTypes,
             targetUserIds = targetUserIds
